Modified opengraph image
I messed with the resolution because Discord wouldn't show the opengraph, but after some time and a reboot it updated after all. It seems like Discord unusually specifically cached the opengraph image as nothing I did to it had any effect. Also corrected the number colors, which is a bug that will be fixed soon.